Start checking why drag proportional to wind speed generates s.s. error
S.S. error is present for both experiment and simulation, so this is not an issue with the simulation model. A standard PID controller will not yield any steady state error either, so this is something solely due to the mc_att_ctrl controller. The error is present even without a motor model in the loop. Changing the controller parameters (integrator limits) does not have any effect on the steady state error. Furthermore, PID gain attenuation and other such optional features are disabled by default. Rate setpoints do not seem to saturate either. It is necessary to go through the controller equations one by one to see where the issue comes from.

Other note: External saturation checks are not implemented in the model. Not required for now but should be implemented later on.