Update the aero model Before, the perpendicular thrust of the rotors of the UAV was incorrectly calculated, as it was rotated by the sideslip angle (which is measured in bodyframe, not motor frame) *before* being rotated from the motor frame to the bodyframe.
Using the AIAA model, each rotor produces two forces: an axial force and a perpendicular force. In order to solve the issue mentioned above, we assume that the 'normal' force is always acting in the *UAV*'s xy-plane, even if the rotors are canted. This means it is not technically 'normal' to the axis of the (canted) rotor, but making this assumption simplifies the implementation a lot (calculating the proper normal force would require using the per-rotor sideslip angle). On the other hand, the axial component of the thrust is still applied axially as this is easy to implement.