PHD-161 Use Uav.C_TAU rather than Uav.B, fixing rate tracking issues
Uav.C_TAU and Uav.B are the same thing, but Uav.B (which is not based on any real value) was 100 times larger. This does not cause issues for planar setups (and in fact helps provide more reaction torque around the z/yaw axis). However, when rotors are tilted it can create a non-insignificant reaction moment around the roll/pitch axes which can saturate the rate integrators, hence leading to steady state error.